+++ extensions/quick-event/src/useCalendar.tsx
@@ -1,5 +1,6 @@
-import { randomId, showToast, ToastStyle } from '@raycast/api';
+import { showToast, Toast, AI, environment } from '@raycast/api';
 import { useState } from 'react';
+import { nanoid } from 'nanoid';
 import { CalendarEvent } from './types';
 import { getEndDate, getStartDate, preprocessQuery } from './dates';
 import osascript from 'osascript-tag';
@@ -13,7 +14,11 @@ export const executeJxa = async (script: string) => {
     if (typeof err === 'string') {
       const message = err.replace('execution error: Error: ', '');
       console.log(err);
-      showToast(ToastStyle.Failure, 'Something went wrong', message);
+      showToast({
+        style: Toast.Style.Failure,
+        title: 'Something went wrong',
+        message: message,
+      });
     }
   }
 };
@@ -35,9 +40,17 @@ export function useCalendar() {
 
         const parsedEvent = Sherlock.parse(preprocessedQuery);
 
+        let location = '';
+        if (environment.canAccess(AI)) {
+          location = await AI.ask(
+            `Help me filter the location for this event: "${query}", just type the location or empty string`,
+          );
+        }
+
         const event: CalendarEvent = {
           ...parsedEvent,
-          id: randomId(),
+          location: location,
+          id: nanoid(),
         };
 
         if (!event.startDate) {
@@ -54,7 +67,11 @@ export function useCalendar() {
       setIsLoading(false);
     } catch (error) {
       console.error('error', error);
-      showToast(ToastStyle.Failure, 'Could not parse event', String(error));
+      showToast({
+        style: Toast.Style.Failure,
+        title: 'Could not parse event',
+        message: String(error),
+      });
     }
   }